Restructuring of functions, added a PCB and inventory listing menu.
With this commit, one can pull a PCB build and check how many components are
available in the database based of the PCB's parts list.
This commit also fixes some of the function calls in both the cli and e7epd class.
Of course this commit also includes some bug fixes and typo fixing as always.